# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DealerSync API",
    "description": "API for DealerSync - Seamless Dealership Operations",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development server" }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login an employee",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "admin@dealersync.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/admin/branches": {
      "post": {
        "summary": "Create a branch (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Downtown Showroom" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Branch created" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/admin/employees": {
      "post": {
        "summary": "Create an employee (Admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Jane Doe" },
                  "email": { "type": "string", "example": "jane@dealersync.com" },
                  "phone": { "type": "string", "example": "9876543210" },
                  "branch_id": { "type": "integer", "example": 1 },
                  "role": { "type": "string", "enum": ["sales", "accounts", "rto", "admin"], "example": "sales" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["name", "email", "phone", "branch_id", "role", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Employee created" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/customers": {
      "post": {
        "summary": "Add a new customer (Sales only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customer_name": { "type": "string", "example": "John Doe" },
                  "phone_number": { "type": "string", "example": "1234567890" },
                  "vehicle": { "type": "string", "example": "Toyota Camry" },
                  "variant": { "type": "string", "example": "LE", "nullable": true },
                  "color": { "type": "string", "example": "Blue", "nullable": true },
                  "price": { "type": "number", "example": 25000.00, "nullable": true }
                },
                "required": ["customer_name", "phone_number", "vehicle"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "customer": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "customer_name": { "type": "string" },
                        "phone_number": { "type": "string" },
                        "vehicle": { "type": "string" },
                        "variant": { "type": "string", "nullable": true },
                        "color": { "type": "string", "nullable": true },
                        "price": { "type": "number", "nullable": true },
                        "created_at": { "type": "string", "format": "date-time" },
                        "created_by": { "type": "integer" },
                        "status": { "type": "string", "enum": ["Pending", "Submitted", "Verified"], "example": "Pending" }
                      }
                    },
                    "uniqueLink": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request" },
          "403": { "description": "Forbidden" }
        }
      },
      "get": {
        "summary": "Fetch customers (Sales see own, Admin see all)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Customers fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "customers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "customer_name": { "type": "string" },
                          "phone_number": { "type": "string" },
                          "vehicle": { "type": "string" },
                          "variant": { "type": "string", "nullable": true },
                          "color": { "type": "string", "nullable": true },
                          "price": { "type": "number", "nullable": true },
                          "created_at": { "type": "string", "format": "date-time" },
                          "created_by": { "type": "integer" },
                          "created_by_name": { "type": "string", "nullable": true },
                          "status": { "type": "string", "enum": ["Pending", "Submitted", "Verified"], "example": "Pending" },
                          "dob": { "type": "string", "format": "date", "nullable": true },
                          "address": { "type": "string", "nullable": true },
                          "mobile_1": { "type": "string", "nullable": true },
                          "mobile_2": { "type": "string", "nullable": true },
                          "email": { "type": "string", "nullable": true },
                          "nominee": { "type": "string", "nullable": true },
                          "nominee_relation": { "type": "string", "nullable": true },
                          "payment_mode": { "type": "string", "nullable": true },
                          "finance_company": { "type": "string", "nullable": true },
                          "finance_amount": { "type": "number", "nullable": true }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "summary": "Fetch single customer details (Public)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 },
            "description": "The ID of the customer"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "customer": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "customer_name": { "type": "string" },
                        "phone_number": { "type": "string" },
                        "vehicle": { "type": "string" },
                        "variant": { "type": "string", "nullable": true },
                        "color": { "type": "string", "nullable": true },
                        "price": { "type": "number", "nullable": true },
                        "created_at": { "type": "string", "format": "date-time" },
                        "status": { "type": "string", "enum": ["Pending", "Submitted", "Verified"], "example": "Pending" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Customer not found" }
        }
      },
      "put": {
        "summary": "Update customer details and status (Customer submits, Sales verifies)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 },
            "description": "The ID of the customer to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["Submitted", "Verified"], "example": "Submitted" },
                  "dob": { "type": "string", "format": "date", "example": "1990-01-01", "nullable": true },
                  "address": { "type": "string", "example": "123 Main St", "nullable": true },
                  "mobile_1": { "type": "string", "example": "1234567890", "nullable": true },
                  "mobile_2": { "type": "string", "example": "0987654321", "nullable": true },
                  "email": { "type": "string", "example": "john.doe@example.com", "nullable": true },
                  "nominee": { "type": "string", "example": "Jane Doe", "nullable": true },
                  "nominee_relation": { "type": "string", "example": "Spouse", "nullable": true },
                  "payment_mode": { "type": "string", "example": "Finance", "nullable": true },
                  "finance_company": { "type": "string", "example": "XYZ Finance", "nullable": true },
                  "finance_amount": { "type": "number", "example": 20000.00, "nullable": true },
                  "aadhar_front": { "type": "string", "format": "binary", "nullable": true },
                  "aadhar_back": { "type": "string", "format": "binary", "nullable": true },
                  "passport_photo": { "type": "string", "format": "binary", "nullable": true }
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "customer": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "customer_name": { "type": "string" },
                        "phone_number": { "type": "string" },
                        "vehicle": { "type": "string" },
                        "variant": { "type": "string", "nullable": true },
                        "color": { "type": "string", "nullable": true },
                        "price": { "type": "number", "nullable": true },
                        "created_at": { "type": "string", "format": "date-time" },
                        "created_by": { "type": "integer" },
                        "status": { "type": "string", "enum": ["Pending", "Submitted", "Verified"], "example": "Submitted" },
                        "dob": { "type": "string", "format": "date", "nullable": true },
                        "address": { "type": "string", "nullable": true },
                        "mobile_1": { "type": "string", "nullable": true },
                        "mobile_2": { "type": "string", "nullable": true },
                        "email": { "type": "string", "nullable": true },
                        "nominee": { "type": "string", "nullable": true },
                        "nominee_relation": { "type": "string", "nullable": true },
                        "payment_mode": { "type": "string", "nullable": true },
                        "finance_company": { "type": "string", "nullable": true },
                        "finance_amount": { "type": "number", "nullable": true }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request - Invalid status or data" },
          "403": { "description": "Forbidden - Unauthorized action" },
          "404": { "description": "Customer not found or not editable" }
        },
        "security": [
          { "bearerAuth": [] }, 
          {} 
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}